% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chess}
\alias{chess}
\title{Chess End-Game - King+Rook versus King+Pawn on a7}
\format{\describe{
\item{\code{TrainX}}{A matrix with 2130 rows and 36 columns.}
\item{\code{TrainY}}{A vector with 2130 rows.}
\item{\code{TestX}}{A matrix with 1066 rows and 36 columns.}
\item{\code{TestY}}{A vector with 1066 rows.}
}}
\usage{
data(chess)
}
\description{
Outcomes of chess games given the board descriptions. \cr
Data set from UCI repository. Training and test splits from SGI.
}
\references{
\link{https://archive.ics.uci.edu/ml/datasets/Chess+(King-Rook+vs.+King-Pawn)}

\link{http://www.sgi.com/tech/mlc/db/chess.names}
}
\keyword{datasets}

