% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{corral}
\alias{corral}
\title{Corral: synthetic data with correlated attributes}
\format{\describe{
\item{\code{TrainX}}{A matrix with 128 rows and 6 columns.}
\item{\code{TrainY}}{A vector with 128 rows.}
}}
\usage{
data(corral)
}
\description{
This is an artificial domain where the target concept is (X1^X2) V (X3^X4). \cr
Data set by R. Kohavi. Training and test splits from SGI.
}
\references{
\link{http://www.sgi.com/tech/mlc/db/corral.names}
}
\keyword{datasets}

