% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{corral_augmented}
\alias{corral_augmented}
\title{Augmented corral: synthetic data with correlated attributes augmented with noise features}
\format{\describe{
\item{\code{TrainX}}{A matrix with 128 rows and 10000 columns.}
\item{\code{TrainY}}{A vector with 128 rows.}
}}
\usage{
data(corral_augmented)
}
\description{
This is an artificial domain where the target concept is (X1^X2) V (X3^X4). \cr
Data set by R. Kohavi. Training and test splits from SGI. \cr
Noise features added by V. Krakovna by shuffling copies of real features.\cr
The SBFC paper uses subsets of this data set with the first 100 and 1000 features.
}
\references{
\href{http://www.sgi.com/tech/mlc/db/corral.names}{SGI listing for corral data set}

\href{arxiv.org/abs/1506.02371}{SBFC paper describing augmentation of corral data set}
}
\keyword{datasets}

