% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{edge_density_plot}
\alias{edge_density_plot}
\title{Plots the density of edges in a given group over the MCMC iterations}
\usage{
edge_density_plot(sbfc_result, group, start = 0, end = 1)
}
\arguments{
\item{sbfc_result}{An object of class \code{sbfc}.}

\item{group}{Which group (0 or 1) to plot edge density for.}

\item{start}{The start of the included range of MCMC iterations (default=0, i.e. starting with the first iteration).}

\item{end}{The end of the included range of MCMC iterations (default=1, i.e. ending with the last iteration).}
}
\description{
Plots the edge density for the given group for a range of the MCMC iterations (indicated by \code{start} and \code{end}).
}
