% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{logpost_plot}
\alias{logpost_plot}
\title{Log posterior trace plot}
\usage{
logpost_plot(sbfc_result, start = 0, end = 1, ...)
}
\description{
Plots the log posterior for a range of the MCMC iterations (indicated by \code{start} and \code{end}).
}

