% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{madelon}
\alias{madelon}
\title{Madelon data set: synthetic data from NIPS 2003 feature selection challenge}
\format{
\describe{
\item{\code{TrainX}}{A matrix with 2000 rows and 500 columns.}
\item{\code{TrainY}}{A vector with 2000 rows.}
\item{\code{TestX}}{A matrix with 600 rows and 500 columns.}
\item{\code{TestY}}{A vector with 600 rows.}
}
}
\usage{
data(madelon)
}
\description{
This is a two-class classification problem. 
The difficulty is that the problem is multivariate and highly non-linear. 
Of the 500 features, 20 are real features, 480 are noise features. \cr
Data set from UCI repository, discretized using median cutoffs.
}
\references{
\href{https://archive.ics.uci.edu/ml/datasets/Madelon}{UCI madelon data set}
}
\keyword{datasets}
