\name{sbfc-package}
\alias{sbfc-package}
\docType{package}
\title{
\packageTitle{sbfc}
}
\description{
\packageDescription{sbfc}
}
\details{
\packageDESCRIPTION{sbfc}
\packageIndices{sbfc}
Run the SBFC algorithm on a data set using the \code{sbfc} function.
Make SBFC graphs based on the MCMC samples using the \code{sbfc_graph} function.
Other analysis, e.g. feature selection plots using \code{signal_var_proportion} (based on how often each variable appeared in the signal group).}
\author{
\packageAuthor{sbfc}
Maintainer: \packageMaintainer{sbfc}
}
\keyword{ package }
