% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{sbfc_graph}
\alias{sbfc_graph}
\title{SBFC graph}
\usage{
sbfc_graph(sbfc_result, iter = 10000, average = T, edge_cutoff = 0.2,
  edges_only = F, ...)
}
\arguments{
\item{sbfc_result}{An object of class \code{sbfc}}

\item{iter}{MCMC iteration of the sampled graph, if \code{average=F} (default=10000)}

\item{average}{Whether to plot an average of sampled MCMC graphs (default=TRUE)}

\item{edge_cutoff}{The average graph includes edges that appear in at least this fraction of the sampled graphs, if \code{average=T} (default=0.2)}

\item{noise_singletons}{Whether to plot single-node trees in the noise group (Group 0), which can be numerous for high-dimensional data sets (default=FALSE).}

\item{names}{Node names for the graph labels (default=\code{c("X1","X2",...)})}
}
\description{
Plots a sampled MCMC graph or an average of sampled graphs.
}
\details{
If the graph renders poorly in the R plot window, try changing the aspect ratio of the plot window.
}

