% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{signal_size_plot}
\alias{signal_size_plot}
\title{Trace plot of Group 1 size}
\usage{
signal_size_plot(sbfc_result, start = 0, end = 1, samples = F)
}
\arguments{
\item{sbfc_result}{An object of class \code{sbfc}.}

\item{start}{The start of the included range of MCMC iterations (default=0, i.e. starting with the first iteration).}

\item{end}{The end of the included range of MCMC iterations (default=1, i.e. ending with the last iteration).}

\item{samples}{Calculate signal group size based on sampled MCMC graphs after burn-in and thinning,
rather than graphs from all iterations (default=FALSE).}
}
\description{
Plots the Group 1 size for a range of the MCMC iterations (indicated by \code{start} and \code{end}).
}
