% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{heart}
\alias{heart}
\title{Heart disease data set: disease outcomes given health attributes}
\format{
\describe{
\item{\code{TrainX}}{A matrix with 270 rows and 13 columns.}
\item{\code{TrainY}}{A vector with 270 rows.}
}
}
\usage{
data(heart)
}
\description{
Data set from UCI repository, discretized using the \code{mdlp} package.
}
\references{
\href{https://archive.ics.uci.edu/ml/datasets/Statlog+(Heart)}{UCI heart data set}
}
\keyword{datasets}
