% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{sbfc_graph}
\alias{sbfc_graph}
\title{SBFC graph}
\usage{
sbfc_graph(sbfc_result, iter = 10000, average = T, edge_cutoff = 0.2,
  single_noise_nodes = F, labels = paste0("X", 1:ncol(sbfc_result$parents)))
}
\arguments{
\item{sbfc_result}{An object of class \code{sbfc}.}

\item{iter}{MCMC iteration of the sampled graph, if \code{average=F} (default=10000).}

\item{average}{Plot an average of sampled MCMC graphs (default=TRUE).}

\item{edge_cutoff}{The average graph includes edges that appear in at least this fraction of the sampled graphs, if \code{average=T} (default=0.2).}

\item{single_noise_nodes}{Plot single-node trees in the noise group (Group 0), which can be numerous for high-dimensional data sets (default=FALSE).}

\item{labels}{Node labels (default=\code{c("X1","X2",...)}).}

\item{edges_only}{Omit single-node trees (default=FALSE).}
}
\description{
Plots a sampled MCMC graph or an average of sampled graphs.
}
\details{
If the graph renders poorly in the R plot window, try changing the aspect ratio of the plot window.
}

